nofv2/requirements.txt
aiohttp
aiohttp_socks
fastapi
numpy
protobuf
python_binance
redis
Requests
ta_lib
uvicorn
